add var check, quiet build

+++ proton-mail-export-cli/run.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 # Usage: ./run.sh /path/to/backup/directory
-podman build -f Containerfile -t proton-mail-export-cli
-podman container run --name proton-mail-export-cli -iv "$1:/backup" proton-mail-export-cli
+if [ -z "${1}" ]; then
+    echo "Please specify a backup directory, i.e. ./run.sh /path/to/backup/directory"
+    exit 1
+fi
+podman build -qf Containerfile -t proton-mail-export-cli
+podman container run --name proton-mail-export-cli -iv "${1}:/backup" proton-mail-export-cli
 podman container rm -f proton-mail-export-cli
 podman image rm proton-mail-export-cli
 podman image prune -af --external
